// schema/dist/propertySchema.d.ts
import z from 'zod';
export declare const propertySchema: z.ZodEffects<z.ZodObject<{
    title: z.ZodString;
    description: z.ZodString;
    bedrooms: z.ZodNumber;
    bathrooms: z.ZodNumber;
    rentalType: z.ZodEnum<["short-term", "long-term", "both"]>;
    pricePerNight: z.ZodOptional<z.ZodNumber>;
    pricePerMonth: z.ZodOptional<z.ZodNumber>;
    depositAmount: z.ZodOptional<z.ZodNumber>;
    address: z.ZodString;
    latitude: z.ZodOptional<z.ZodNumber>;
    longitude: z.ZodOptional<z.ZodNumber>;
    amenities: z.ZodArray<z.ZodString, "many">;
    availability: z.ZodOptional<z.ZodAny>;
    maxGuests: z.ZodNumber;
}, "strip", z.ZodTypeAny, {
    title: string;
    description: string;
    bedrooms: number;
    bathrooms: number;
    rentalType: "short-term" | "long-term" | "both";
    address: string;
    amenities: string[];
    maxGuests: number;
    pricePerNight?: number | undefined;
    pricePerMonth?: number | undefined;
    depositAmount?: number | undefined;
    latitude?: number | undefined;
    longitude?: number | undefined;
    availability?: any;
}, {
    title: string;
    description: string;
    bedrooms: number;
    bathrooms: number;
    rentalType: "short-term" | "long-term" | "both";
    address: string;
    amenities: string[];
    maxGuests: number;
    pricePerNight?: number | undefined;
    pricePerMonth?: number | undefined;
    depositAmount?: number | undefined;
    latitude?: number | undefined;
    longitude?: number | undefined;
    availability?: any;
}>, {
    title: string;
    description: string;
    bedrooms: number;
    bathrooms: number;
    rentalType: "short-term" | "long-term" | "both";
    address: string;
    amenities: string[];
    maxGuests: number;
    pricePerNight?: number | undefined;
    pricePerMonth?: number | undefined;
    depositAmount?: number | undefined;
    latitude?: number | undefined;
    longitude?: number | undefined;
    availability?: any;
}, {
    title: string;
    description: string;
    bedrooms: number;
    bathrooms: number;
    rentalType: "short-term" | "long-term" | "both";
    address: string;
    amenities: string[];
    maxGuests: number;
    pricePerNight?: number | undefined;
    pricePerMonth?: number | undefined;
    depositAmount?: number | undefined;
    latitude?: number | undefined;
    longitude?: number | undefined;
    availability?: any;
}>;
export type PropertySchema = z.infer<typeof propertySchema>;